networkmanager: Fix sysvinit service conflict
When using the new networkmanager recipe (networkmanager-daemon
package), both a NetworkManager.service systemd unit and a
network-manager sysvinit script are packaged in the image. When booting
a system with systemd, systemd-sysv-generator will generate an
early/boot-time .service file for the sysv init script since its name
does not match the systemd unit. This results in a conflict between
essentially identical services for the NetworkManager daemon.

To work-around this, create a .bbappend with a symlink for the systemd
unit that matches the sysv init script name. This is non-invasive and
will simply alias the unit and prevent a duplicate early unit from being
created at system boot.

Signed-off-by: Ken Sloat <ken.s@variscite.com>